requirements.txt
torch>=2.0.0
torchvision
einops
transformers>=4.25.1
diffusers>=0.15.0
accelerate
safetensors
pillow
av
huggingface_hub
imageio
